requirements.txt
bibtexparser>=2.0.0b1
pyyaml
